feat(diagnostics): put diagnostics on ther column
The current implementation of the diagnostics, when enabling the
`virtual_text` option was to put all virtual text extmarks on column 0,
no matter the actual `vim.Diagnostic` data.

This commit allows to put those extmarks on the 'good' column, by
looking at the diagnostic's `end_col` field and using that for the
extmark position.